requirements.txt
ultralytics==8.3.208
streamlit==1.28.0
opencv-python-headless==4.8.1.78
Pillow==10.0.1
numpy==1.24.3
torch>=2.0.0
pandas==2.0.3